Put regs, selectors, metrics into own source files
